金额大写方法抽出
