+++ src/components/Typefield/Typefield.vue
@@ -27,54 +27,11 @@
   </div>
 </template>
 <script>
-import {oneOf,formatnumber} from '../../util/tools'
+import {oneOf,formatnumber,changeTipsVal,cutNum,divideNum} from '../../util/tools'
 import Emitter from '../../mixins/emitter';
 import Locale from '../../mixins/locale';
 import Drop from '../Select/Dropdown.vue'
 import TransferDom from '../../directives/transfer-dom';
-Number.prototype.toFixedSelf = function (n) {
-  if (n > 20 || n < 0) {
-      throw new RangeError('toFixed() digits argument must be between 0 and 20');
-  }
-  const number = this;
-  if (isNaN(number) || number >= Math.pow(10, 21)) {
-      return number.toString();
-  }
-  if (typeof (n) == 'undefined' || n == 0) {
-      return (Math.round(number)).toString();
-  }
-  let result = number.toString();
-  const arr = result.split('.');
-  // 整数的情况
-  if (arr.length < 2) {
-      result += '.';
-      for (let i = 0; i < n; i += 1) {
-          result += '0';
-      }
-      return result;
-  }
-  const integer = arr[0];
-  const decimal = arr[1];
-  if (decimal.length == n) {
-      return result;
-  }
-  if (decimal.length < n) {
-      for (let i = 0; i < n - decimal.length; i += 1) {
-          result += '0';
-      }
-      return result;
-  }
-  result = integer + '.' + decimal.substr(0, n);
-  const last = decimal.substr(n, 1);
-
-  // 四舍五入，转换为整数再处理，避免浮点数精度的损失
-  if (parseInt(last, 10) >= 5) {
-      const x = Math.pow(10, n);
-      result = (Math.round((parseFloat(result) * x)) + 1) / x;
-      result = result.toFixedSelf(n);
-  }
-  return result;
-};
 const prefixCls = 'h-typefield';
 export default {
   name: 'Typefield',
@@ -222,7 +179,7 @@ export default {
       }
     },
     inputValue(val){
-      this.viewValue=this.changeTipsVal(val)
+      this.viewValue=changeTipsVal(val,this.integerNum,this.suffixNum,this.isround)
     },
     tipShow(val){
       if (val&&this.transfer) {
@@ -250,9 +207,9 @@ export default {
       }
       let val = this.inputValue?this.inputValue.trim():this.inputValue;
       if (val && val!="") {
-        val = this.notFormat?e.target.value:this.formatNum(e.target.value.trim().replace(/,/g,''));
+        val = this.notFormat?e.target.value:this.formatNum(e.target.value.trim().replace(/,/g,''),this.integerNum,this.suffixNum);
         if ((this.divided || this.immeDivided) && this.type=="money") {
-          this.inputValue = this.divideNum(val);
+          this.inputValue = divideNum(val);
         }else{
           this.inputValue = val;
         }
@@ -334,40 +291,12 @@ export default {
       if(!this.immeDivided){
         this.inputValue = value;
       }else{
-        this.inputValue = this.divideNum(value);
+        this.inputValue = divideNum(value);
       }
       this.bigShow(this.type,value);
       this.$emit('input', value);
       this.$emit('on-keyup', value);
       this.dispatch('FormItem', 'on-form-change', value);
-    },
-    divideNum(num){
-      let revalue="";
-      let array=String(num).split(".");
-      let pointStr = array[1]?'.'+array[1]:''
-      array[0] = array[0].replace(/-/g, "")
-      if(array[0].length>3){
-        while(array[0].length>3){
-          revalue=","+array[0].substring(array[0].length-3,array[0].length)+revalue;
-          array[0]=array[0].substring(0,array[0].length-3);
-        }
-      }
-      return num>=0?array[0]+revalue+pointStr:'-'+array[0]+revalue+pointStr;
-    },
-    changeTipsVal (value){
-      value  = String(value).replace(/[^0-9\.-]/g,"");
-      var firstChar = value.substring(0,1);
-      value = this.cutNum(value,this.integerNum);
-      if(value.split(".")[1] && value.split(".")[1].length > 2){
-        var isround = this.isround;
-        if(isround&&isround==true){
-          value = parseFloat(value).toFixedSelf(this.suffixNum);
-        }else{
-          var suf = value.split(".")[1].substr(0,this.suffixNum);
-          value = value.split(".")[0]+"."+suf;
-        }
-      }
-      return this.numtochinese(value + "",this.suffixNum);
     },
     // 转换科学技术法为数值---Number(val)会在小数点7位后以科学计数法返回，影响格式化
     changeRexNum (val) {
@@ -423,10 +352,8 @@ export default {
       }
       return result;
     },
-    formatNum (value){
+    formatNum (value,integerNum,suffixNum){
       value = value.trim().replace(/,/g, "");
-      var suffixNum = this.suffixNum;
-      var integerNum = this.integerNum;
       value  = value.replace(/[^0-9\.-]/g,"")||'';
       var firstChar = value.substring(0,1)||'';
       if(this.type == "cardNo"){
@@ -446,7 +373,7 @@ export default {
         var intLength = valArr.length>0?valArr[0].length:value;
         if (integerNum<16 || intLength<16) {
           value=value.replace("-", "")
-          value = this.cutNum(value,this.integerNum);
+          value = cutNum(value,integerNum);
           if (value=='') return;
           if (this.isround) {
             // 7位小数Number会转成科学计数法
@@ -507,226 +434,6 @@ export default {
       value = val1+curVal2;
       return value;
     },
-    numtochinese (Num,suffixNumber) {
-      for (var i = Num.length - 1; i >= 0; i--) {
-        Num = Num.replace(",", "")// 替换tomoney()中的“,”
-        Num = Num.replace(" ", "")// 替换tomoney()中的空格
-      }
-      Num = Num.replace("￥", "")// 替换掉可能出现的￥字符
-      if (isNaN(Num)) {
-        // 验证输入的字符是否为数字
-        // alert("请检查小写金额是否正确");
-        return;
-      }
-      // ---字符处理完毕，开始转换，转换采用前后两部分分别转换---//
-      var part = String(Num).split(".");
-      var newchar = "";
-      // 小数点前进行转化
-      for (i = part[0].length - 1; i >= 0; i--) {
-        if (part[0].length > 17) {
-        //   alert("");
-          return "位数过大，无法计算";
-        }// 若数量超过拾亿单位，提示
-        var tmpnewchar = ""
-        var perchar = part[0].charAt(i);
-        switch (perchar) {
-        case "0":
-          tmpnewchar = "零" + tmpnewchar;
-          break;
-        case "1":
-          tmpnewchar = "壹" + tmpnewchar;
-          break;
-        case "2":
-          tmpnewchar = "贰" + tmpnewchar;
-          break;
-        case "3":
-          tmpnewchar = "叁" + tmpnewchar;
-          break;
-        case "4":
-          tmpnewchar = "肆" + tmpnewchar;
-          break;
-        case "5":
-          tmpnewchar = "伍" + tmpnewchar;
-          break;
-        case "6":
-          tmpnewchar = "陆" + tmpnewchar;
-          break;
-        case "7":
-          tmpnewchar = "柒" + tmpnewchar;
-          break;
-        case "8":
-          tmpnewchar = "捌" + tmpnewchar;
-          break;
-        case "9":
-          tmpnewchar = "玖" + tmpnewchar;
-          break;
-        }
-        switch (part[0].length - i - 1) {
-        case 0:
-          tmpnewchar = tmpnewchar + "元";
-          break;
-        case 1:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "拾";
-          break;
-        case 2:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "佰";
-          break;
-        case 3:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "仟";
-          break;
-        case 4:
-          tmpnewchar = tmpnewchar + "万";
-          break;
-        case 5:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "拾";
-          break;
-        case 6:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "佰";
-          break;
-        case 7:
-          if (perchar != 0)
-            tmpnewchar = tmpnewchar + "仟";
-          break;
-        case 8:
-          tmpnewchar = tmpnewchar + "亿";
-          break;
-        case 9:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "拾";
-          break;
-        case 10:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "百";
-          break;
-        case 11:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "仟";
-          break;
-        case 12:
-          tmpnewchar = tmpnewchar + "兆";
-          break;
-        case 13:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "拾";
-          break;
-        case 14:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "百";
-          break;
-        case 15:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "仟";
-          break;
-        case 16:
-          if (perchar != 0)
-          tmpnewchar = tmpnewchar + "京";
-          break;
-        case 17:
-          tmpnewchar = tmpnewchar + "拾";
-          break;
-        }
-        newchar = tmpnewchar + newchar;
-      }
-      // 小数点之后进行转化
-      if (Num.indexOf(".") != -1) {
-        if (part[1].length > 2) {
-          //alert("小数点之后只能保留两位,系统将自动截段");
-          var tempNum = parseFloat(Num);
-          Num = tempNum.toFixedSelf(suffixNumber);
-          part = String(Num).split(".");
-        }
-        for (i = 0; i < part[1].length; i++) {
-          tmpnewchar = ""
-          perchar = part[1].charAt(i)
-          switch (perchar) {
-          case "0":
-            tmpnewchar = "零" + tmpnewchar;
-            break;
-          case "1":
-            tmpnewchar = "壹" + tmpnewchar;
-            break;
-          case "2":
-            tmpnewchar = "贰" + tmpnewchar;
-            break;
-          case "3":
-            tmpnewchar = "叁" + tmpnewchar;
-            break;
-          case "4":
-            tmpnewchar = "肆" + tmpnewchar;
-            break;
-          case "5":
-            tmpnewchar = "伍" + tmpnewchar;
-            break;
-          case "6":
-            tmpnewchar = "陆" + tmpnewchar;
-            break;
-          case "7":
-            tmpnewchar = "柒" + tmpnewchar;
-            break;
-          case "8":
-            tmpnewchar = "捌" + tmpnewchar;
-            break;
-          case "9":
-            tmpnewchar = "玖" + tmpnewchar;
-            break;
-          }
-          if (i == 0)
-            tmpnewchar = tmpnewchar + "角";
-          if (i == 1)
-            tmpnewchar = tmpnewchar + "分";
-          newchar = newchar + tmpnewchar;
-        }
-      }
-      //替换所有无用汉字
-      while (newchar.search("零零") != -1)
-        newchar = newchar.replace("零零", "零");
-        newchar = newchar.replace("零亿", "亿");
-        newchar = newchar.replace("亿万", "亿");
-        newchar = newchar.replace("零万", "万");
-        newchar = newchar.replace("零元", "元");
-        newchar = newchar.replace("零角", "");
-        newchar = newchar.replace("零分", "");
-
-        newchar = newchar.replace("亿万", "亿");
-        newchar = newchar.replace("兆亿", "兆");
-        newchar = newchar.replace("零兆", "兆");
-        newchar = newchar.replace("京兆", "京");
-
-      if (newchar.charAt(newchar.length - 1) == "元"
-          || newchar.charAt(newchar.length - 1) == "角")
-        newchar = newchar + "整";
-
-      var digit = ['壹', '贰', '叁', '肆', '伍', '陆', '柒', '捌', '玖'];
-      var _i = 0
-      while(newchar.length > 0){
-        if(digit.indexOf(newchar[0]) < 0){
-          newchar = newchar.substr(1);
-        }else{
-          break;
-        }
-      }
-
-      var firstChar = Num.substring(0,1);
-      if(firstChar=="-"){
-        newchar = "负"+newchar;
-      }
-
-      var lastChar=newchar.charAt(newchar.length-1);
-      if("零" == lastChar){
-        newchar=newchar.substring(0,newchar.length-1);
-        newchar=newchar+"整";
-      }
-
-      if(parseFloat(Num)==0){
-        newchar="零元整";
-      }
-      return newchar;
-    },
     fillZero(number, bitNum) {
       /// 小数位不够，用0补足位数
       var f_x = parseFloat(number);
@@ -747,18 +454,6 @@ export default {
       }
       return s_x;
     },
-    cutNum(value,integerNum){
-      var arrNum=value.split(".");
-      if(arrNum.length>0){
-        var integerNumber=arrNum[0].substr(0,this.integerNum);
-        if(arrNum.length>1){
-          value=integerNumber+"."+arrNum[1]
-        }else{
-          value=integerNumber
-        }
-      }
-      return value;
-    },
     setNum(value,suffixNum,integerNum){
       if (isNaN(value)) return ;
       if (suffixNum>0) {
@@ -773,7 +468,7 @@ export default {
     bigShow(type,value){
       if (type=='money'&&value) {
         if (this.bigTips) {
-          this.bigNum = this.changeTipsVal(value);
+          this.bigNum = changeTipsVal(value,this.integerNum,this.suffixNum,this.isround);
           this.tipShow = Boolean(this.bigNum);
         }
       }else{
@@ -794,11 +489,11 @@ export default {
         formatVal = this.inputValue;
       } else {
         // 失焦的时候才格式化，避免不能增删小数位的问题
-        formatVal = this.notFormat || this.havefocused || !val ? val : this.formatNum(val.replace(/,/g, ''));
+        formatVal = this.notFormat || this.havefocused || !val ? val : this.formatNum(val.replace(/,/g, ''),this.integerNum,this.suffixNum);
         if (formatVal && this.divided && this.type=='money' && !this.havefocused) {
-          this.inputValue = this.divideNum(formatVal);
+          this.inputValue = divideNum(formatVal);
         } else if (this.immeDivided && formatVal && this.type=='money'){
-          this.inputValue = this.divideNum(formatVal);
+          this.inputValue = divideNum(formatVal);
         } else {
           this.inputValue = formatVal;
         }
